Revert "#163842014 Refactored code file structure "
